Add onboarding modal and hero banner
