+++ src/components/index.ts
@@ -1,0 +1,2 @@
+export { default as OnboardForm } from "./OnboardForm";
+export { default as Modal } from "./ui/modal";
